Move error types into their own files (#9199)
